# packages/api-client/openapi.json
{
    "components": {
        "schemas": {
...
            "CreateDocumentRequest": {
                "properties": {
                    "description": {
                        "maxLength": 5000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "initialVersionLabel": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "name": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "slug": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "visibility": {
                        "pattern": "public|unlisted|private",
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
            "CreateVersionRequest": {
                "properties": {
                    "cloneFromVersionId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "versionLabel": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    }
                },
                "required": [
                    "versionLabel"
                ],
                "type": "object"
            },
            "DocumentResponse": {
                "properties": {
                    "branding": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "createdBy": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "currentVersionId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "publishedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "visibility": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateDocumentRequest": {
                "properties": {
                    "branding": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "description": {
                        "maxLength": 5000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "name": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "slug": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "visibility": {
                        "pattern": "public|unlisted|private",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateVersionRequest": {
                "properties": {
                    "changelogMd": {
                        "maxLength": 50000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "versionLabel": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "VersionResponse": {
                "properties": {
                    "changelogMd": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "documentId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "publishedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "versionLabel": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/v1/documents": {
            "get": {
                "operationId": "listDocuments",
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-controller"
                ]
            },
            "post": {
                "operationId": "createDocument",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDocumentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "tags": [
                    "document-controller"
                ]
            }
        },
        "/api/v1/documents/{documentId}": {
            "delete": {
                "operationId": "deleteDocument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "tags": [
                    "document-controller"
                ]
            },
            "get": {
                "operationId": "getDocument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-controller"
                ]
            },
            "patch": {
                "operationId": "updateDocument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDocumentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-controller"
                ]
            }
        },
        "/api/v1/documents/{documentId}/versions": {
            "get": {
                "operationId": "listVersions",
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/VersionResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-version-controller"
                ]
            },
            "post": {
                "operationId": "createVersion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "documentId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateVersionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionResponse"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "tags": [
                    "document-version-controller"
                ]
            }
        },
...
        "/api/v1/versions/{versionId}": {
            "get": {
                "operationId": "getVersion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "versionId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-version-controller"
                ]
            },
            "patch": {
                "operationId": "updateVersion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "versionId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVersionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-version-controller"
                ]
            }
        },
        "/api/v1/versions/{versionId}/archive": {
            "post": {
                "operationId": "archiveVersion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "versionId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-version-controller"
                ]
            }
        },
...
        "/api/v1/versions/{versionId}/publish": {
            "post": {
                "operationId": "publishVersion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "versionId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "document-version-controller"
                ]
            }
        }
    },
    "servers": [
        {
            "description": "Generated server url",
            "url": "http://localhost:8080"
        }
    ]
}